feat: Add tap gesture to dismiss keyboard on InitialScreen
This commit introduces a tap gesture on the `InitialScreen` that allows users to dismiss the keyboard by tapping anywhere outside the input fields. The `_onScaffoldTap` function has been added to handle the gesture and unfocus any active input fields, enhancing the user experience by providing a convenient way to close the keyboard.